Moved login and register to components
