+++ client/src/AppRouter.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import {BrowserRouter, Switch, Route} from 'react-router-dom'
 import App from './App'
-import Login from './Login'
-import Register from './Register'
+import Login from './components/Login'
+import Register from './components/Register'
 
 const AppRouter = () => (
     <BrowserRouter>
